iOS NUI: for main doc view, move render control into the page controller
All our current document views use continuous scrolling and we have
the ARDKViewRenderer class to control the rendering. For the main doc view
we can substitute a different page controller, for which that mode of
render control may not be appropriate. This commit moves the responsibility
for controlling render into the page controller. That's only for the main
view. The pertaining mechanism is still appropriate for the other views.

<log></log>