iOS NUI: implement page scrolling in book mode
Wrap the UIPageViewController within an outer UIScrollView and an inner
"content" UIView. Constraints control the size of the inner UIView so that
scrolling is active only when focussing on the left or right page. In two
up mode, the content view is matched in size to the UIScrolView so that
no scrolling occurs.

<log></log>