iOS NUI: change the mechanism for queuing renders in book mode
The previous mechanism was specifically coded for the render call. This
new approach defines a general mechanism for queuing asynchonous blocks.
Use of the new mechanism will allow future changes to the render call,
with additional parameters without having to also alter the queue to
carry the new parameters.

We also change the queuing strategy to drop all but the currently running
action when queuing a new action, so as to avoid having more than one
waiting.

<log></log>