iOS NUI: handle zoomed in rendering in book mode
For the two up mode, an individual bitmap is allocated to the two visible
pages and two either side, so that the either-side pages can be
pre-rendered and available to take part in page-turning annimations.

In one-up mode, a different strategy is required. The page is very much
bigger and we can afford to render only what is visible and because of
that we have to repeatedly render as the page is scrolled. That mode of
rendering requires two page-sized bitmaps, which are used in alternation.
Luckily, we no longer need to render pages to either side, so we reuse the
bitmaps to either side. We have four such bitmaps, rather than two, and
they are half the size they need to be, but we allocate them in such a way
that they can be used in pairs (the pairs are actually allocated as part
of a single larger bitmap). We retain the bitmap used for two up rendering
for use as a low-res background (although that feature isn't complete in
this commit).

<log></log>