iOS NUI: provide for adjustment to newly added signature fields
Previously, when a signature field was added, it's position was initially
adjustable, but there was no way to delete it and once the user clicked
away to deselect it, there was no way to alter it at all. This commit
allows just the signature fields that have been added in the current
session to be repositioned or deleted: the dialog that pops up when the
field is tapped, now includes those options in addition to an option for
signing it.

<log></log>