iOS NUI: fix confusing behaviour of signature-field highlighting
Form fields are hightlighted by default, with the hightlight removed for
just the currently focussed field, usually the one being filled. Signature
fields were being left focussed and hence not highlighted after signing
and verification, when the lack of a highlight might be confusing to the
user. This commit, drops focus after signature operations to avoid that
behaviour.

<log></log>