iOS NUI: fix possible exception in UITextInput implementation
This fixes bug 703132.

Our implementation of UITextInput's textInRange didn't account for the case
of being asked for range that didn't fit within the current text. We mostly
got away with this because the text's being out of sync with the OS was a
rare occurrence. One of the few occasions was when a form field's
validation failed and caused the field to be emptied, and that case only
when part of backgrounding the app. This commit protects the call in the
general case.

<log></log>