iOS NUI: avoid fixed height in certificate picker UI
This fixes Bug 703131

The stack view that contains the list of cerificates has a fixed height
set within the storey board, although the height will be determined from
content once populated. We have left the constraint present for the sake
of the case where there are no certificates, but set the priority very
low so that content takes priority.

The bug wasn't reproducible on all devices. iPad Pro 4th Generation is pne
for which it does.

<log></log>