iOS MuPDF: prevent some cases of file corruption causing crashes
If a PDF file is broken then the mupdf core may throw when a page of a
document is requested. We were handling that and recording a NULL value
in our page cache but, with redaction, came new code that didn't test for
NULL. The most important part of this commit is to add a test for NULL to
that redaction code. Also, while investigating this problem, it was noticed
that, through use of an array to hold the page cache, we'd introduced an
algorithm for page loading that was n-squared in the number of pages. The
second part of this commit addresses that by using a dictionary keyed by
page number in place of the array.

<log></log>