iOS NUI: partial revert of 2cfc6ef54938c69ce617b9f89c1511abca769bc9
The referenced commit fixed a crash, but also included an optimisation.
There was a slip in the optimisation. Rather than fix the slip, it seems
safer to roll it back on this release branch, leaving just the crash fix.
A fix for the optimisation part will be added to the master branch.

<log></log>