iOS NUI: fix AppKit samples build under Xcode 12.3
This is a workaround for bug 703289. Xcode 12.3 fails to build the sample,
giving error: "Building for iOS, but the linked and embedded framework
‘Structure.framework’ was built for iOS + iOS Simulator". These changes
of settings allows the build to work, not entirely sure why. We have a
better solution in the pipeline that changes the building of the
framework and makes the sample setting change unnecessary.

<log></log>